fix: update logic checking whether mypy flags passed
Logic added in f1405064bec2956e7690c2f5ab3e029007ed893f failed to check
for the case of an empty list being passed as the mypy flags, resulting
in _no_ flags being passed to mypy.

This update just ensures that if the mypy_flags var is '[]' (empty list)
then mypy will be called with the default of --strict as
before/expected.